feat: 커스텀 permission 만들기

+++ lion/posts/views.py
@@ -14,7 +14,7 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from django.http import Http404
-from rest_framework.permissions import IsAuthenticatedOrReadOnly # jwt 세션
+from config.permissions import IsAuthenticatedUserAndAllowTime
 
 class PostList(APIView):
     def post(self, request, format=None):
@@ -30,7 +30,7 @@ class PostList(APIView):
         return Response(serializer.data)
     
 class PostDetail(APIView):
-    permission_classes = [IsAuthenticatedOrReadOnly]
+    permission_classes = [IsAuthenticatedUserAndAllowTime]
 
     def get(self, request, post_id):
         post = get_object_or_404(Post, id=post_id)
@@ -39,6 +39,7 @@ class PostDetail(APIView):
     
     def put(self, request, post_id):
         post = get_object_or_404(Post, id=post_id)
+        self.check_object_permissions(request, post)
         serializer = PostSerializer(post, data=request.data)
         if serializer.is_valid(): # update이니까 유효성 검사 필요
             serializer.save()
@@ -47,6 +48,7 @@ class PostDetail(APIView):
     
     def delete(self, request, post_id):
         post = get_object_or_404(Post, id=post_id)
+        self.check_object_permissions(request, post)
         post.delete()
         return Response(
 	        {
